+++ csvkit/utilities/csvgrep.py
@@ -1,9 +1,11 @@
 #!/usr/bin/env python
 
 import re
+import sys
 from argparse import FileType
 
 import agate
+import six
 
 from csvkit.cli import CSVKitUtility
 from csvkit.grep import FilteringCSVReader
@@ -14,23 +16,35 @@ class CSVGrep(CSVKitUtility):
     override_flags = ['L', 'blanks', 'date-format', 'datetime-format']
 
     def add_arguments(self):
+        # I feel that there ought to be a better way to do this across Python 2 and 3.
+        def option_parser(bytestring):
+            if six.PY2:
+                return bytestring.decode(sys.getfilesystemencoding())
+            else:
+                return bytestring
+
         self.argparser.add_argument('-n', '--names', dest='names_only', action='store_true',
                                     help='Display column names and indices from the input CSV and exit.')
         self.argparser.add_argument('-c', '--columns', dest='columns',
-                                    help='A comma separated list of column indices or names to be searched.')
-        self.argparser.add_argument('-m', '--match', dest="pattern", action='store',
+                                    help='A comma separated list of column indices, names or ranges to be searched, e.g. "1,id,3-5".')
+        self.argparser.add_argument('-m', '--match', dest="pattern", action='store', type=option_parser,
                                     help='The string to search for.')
-        self.argparser.add_argument('-r', '--regex', dest='regex', action='store',
+        self.argparser.add_argument('-r', '--regex', dest='regex', action='store', type=option_parser,
                                     help='If specified, must be followed by a regular expression which will be tested against the specified columns.')
         self.argparser.add_argument('-f', '--file', dest='matchfile', type=FileType('r'), action='store',
                                     help='If specified, must be the path to a file. For each tested row, if any line in the file (stripped of line separators) is an exact match for the cell value, the row will pass.')
         self.argparser.add_argument('-i', '--invert-match', dest='inverse', action='store_true',
                                     help='If specified, select non-matching instead of matching rows.')
+        self.argparser.add_argument('-a', '--any-match', dest='any_match', action='store_true',
+                                    help='If specified, select rows where any column matches instead of all columns.')
 
     def main(self):
         if self.args.names_only:
             self.print_column_names()
             return
+
+        if self.additional_input_expected():
+            sys.stderr.write('No input file or piped data provided. Waiting for standard input:\n')
 
         if not self.args.columns:
             self.argparser.error('You must specify at least one column to search using the -c option.')
@@ -50,6 +64,7 @@ class CSVGrep(CSVKitUtility):
             pattern = re.compile(self.args.regex)
         elif self.args.matchfile:
             lines = set(line.rstrip() for line in self.args.matchfile)
+            self.args.matchfile.close()
 
             def pattern(x):
                 return x in lines
@@ -57,7 +72,7 @@ class CSVGrep(CSVKitUtility):
             pattern = self.args.pattern
 
         patterns = dict((column_id, pattern) for column_id in column_ids)
-        filter_reader = FilteringCSVReader(rows, header=False, patterns=patterns, inverse=self.args.inverse)
+        filter_reader = FilteringCSVReader(rows, header=False, patterns=patterns, inverse=self.args.inverse, any_match=self.args.any_match)
 
         output = agate.csv.writer(self.output_file, **writer_kwargs)
         output.writerow(column_names)