+++ csvkit/utilities/csvlook.py
@@ -7,7 +7,6 @@ from csvkit.cli import CSVKitUtility
 
 class CSVLook(CSVKitUtility):
     description = 'Render a CSV file in the console as a Markdown-compatible, fixed-width table.'
-    buffers_input = True
 
     def add_arguments(self):
         self.argparser.add_argument('--max-rows', dest='max_rows', type=int,
@@ -22,6 +21,9 @@ class CSVLook(CSVKitUtility):
                                     help='Disable type inference when parsing the input.')
 
     def main(self):
+        if self.additional_input_expected():
+            self.argparser.error('You must provide an input file or piped data.')
+
         table = agate.Table.from_csv(
             self.input_file,
             skip_lines=self.args.skip_lines,